requirements.txt
playwright==1.40.0
gspread==6.0.0
google-auth==2.25.0
APScheduler==3.10.4
flask==3.0.0
